Create a dedicated module for help tests.
